png: encode 16-bit Rec.2020 / PQ (BT.2100); cICP-signalled, output only
canvas_write_png now emits a BT.2100 PNG -- 16-bit, Rec.2020 primaries, PQ
(ST 2084) transfer, a cICP chunk -- regardless of working space.  The pixel
pipeline (canvas_encode_png / surface_to_pq16) takes the premultiplied f16
surface to linear sRGB light (an sRGB surface decodes; a linear surface is
already linear, extended HDR / wide-gamut values intact), rotates to Rec.2020
primaries, scales to the PQ range against the reference white, PQ-encodes, and
quantizes to 16-bit.  Reference white is CNVS_REF_WHITE_NITS (203, BT.2408; one
knob).  Wide gamut rides the existing extended-linear surface; no parameterized
primaries needed (a fixed sRGB->Rec.2020 matrix at output).

The Rec.2020 matrix and PQ transfer join cnvs_color (cnvs_linear_srgb_to_rec2020,
cnvs_pq_oetf), checked against a double-precision oracle on the spec's canonical
fractions plus white/primary known answers.  The cnvs_png encoder takes 16-bit
samples (big-endian, Up-filtered); bench/fuzz feed uint16.  The byte gate
re-encodes the replayed canvas through canvas_encode_png and compares file bytes.
All 39 gallery PNGs regenerated as 16-bit PQ; coverage and docs updated.

+++ src/cnvs_png.h
@@ -3,22 +3,24 @@
 #include <ptrcheck.h>
 #include <stdint.h>
 
-// PNG encode: 8-bit RGBA (color type 6), non-interlaced, every row Up-filtered
-// (filter byte 2), one IDAT carrying a cnvs_zlib deflate stream.  Up-only is a
-// deliberate trade (see cnvs_png.c): it vectorizes as whole-row subtract/add
-// with no left-neighbor recurrence, and the adaptive per-row filter chooser
-// only beat it by ~2% across the gallery corpus.  Deterministic: identical
-// pixels -> identical PNG bytes, always.
+// PNG encode: 16-bit RGBA (colour type 6), non-interlaced, every row Up-filtered
+// (filter byte 2), one IDAT carrying a cnvs_zlib deflate stream, plus a cICP
+// chunk signalling BT.2100 (Rec.2020 primaries, PQ transfer).  The pixels are
+// already in that encoding; the colour pipeline that produces them is in
+// canvas.c (canvas_encode_png).  Up-only is a deliberate trade (see cnvs_png.c):
+// it vectorizes as a whole-row subtract with no left-neighbor recurrence.
+// Deterministic: identical pixels -> identical PNG bytes, always.
 
 // Encode to a malloc'd buffer (caller frees), storing its byte length in
-// *outlen.  `pixels` is tightly packed RGBA8, top row first.  On bad dimensions
-// (non-positive, > 16384 -- bounded so every size computation fits an int) or
-// out of memory: NULL, *outlen = 0.
+// *outlen.  `pixels` is tightly packed RGBA, top row first, four uint16 samples
+// per pixel (native-endian; serialized big-endian here).  On bad dimensions
+// (non-positive, or large enough that the deflate size arithmetic would not fit
+// an int) or out of memory: NULL, *outlen = 0.
 uint8_t *__counted_by_or_null(*outlen)
-cnvs_png_encode(uint8_t const *__counted_by(width * height * 4) pixels,
+cnvs_png_encode(uint16_t const *__counted_by(width * height * 4) pixels,
                 int width, int height, int *__single outlen);
 
 // cnvs_png_encode + write the bytes to `path`.
 bool cnvs_png_write(char const *__null_terminated path,
-                    uint8_t const *__counted_by(width * height * 4) pixels,
+                    uint16_t const *__counted_by(width * height * 4) pixels,
                     int width, int height);
